requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
groq==0.4.0
python-dotenv==1.0.0
pocket-tts==0.1.0
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
huggingface-hub==0.19.4